elt/pyproject.toml
[project]
name = "elt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "clickhouse-connect>=0.10.0",
    "duckdb>=1.4.0",
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
